+++ src/Homie/Limits.hpp
@@ -13,7 +13,7 @@
 namespace HomieInternals {
   // max setting elements
   // 6 elements at root, 9 elements at wifi, 6 elements at mqtt, 1 element at ota, max settings elements
-  const uint16_t MAX_JSON_CONFIG_ARDUINOJSON_BUFFER_SIZE = JSON_OBJECT_SIZE(6) + JSON_OBJECT_SIZE(9) + JSON_OBJECT_SIZE(6) + JSON_OBJECT_SIZE(1) + JSON_OBJECT_SIZE(MAX_CONFIG_SETTING_SIZE);
+  const uint16_t MAX_JSON_CONFIG_ARDUINOJSON_BUFFER_SIZE = JSON_OBJECT_SIZE(6) + JSON_OBJECT_SIZE(9) + JSON_OBJECT_SIZE(6) + JSON_OBJECT_SIZE(1) + JSON_OBJECT_SIZE(uint8_t(MAX_CONFIG_SETTING_SIZE));
 
   const uint8_t MAX_MAC_LENGTH = 6;
   const uint8_t MAX_MAC_STRING_LENGTH = (MAX_MAC_LENGTH * 2) + 1;
